refactor(keeper): splitting up keeper into seperate files & removing rest client
