--- x/inter-tx/client/rest/rest.go
@@ -1,12 +1,0 @@
-package rest
-
-import (
-	"fmt"
-	//	"github.com/cosmos/cosmos-sdk/client/context"
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/gorilla/mux"
-)
-
-func RegisterRoutes(cliCtx client.Context, r *mux.Router) {
-	fmt.Println("TODO")
-}
